Convert difference to int

+++ encarne/client/commands.py
@@ -143,7 +143,7 @@ def execute_run(args):
                     copy = False
                 else:
                     difference = origin_filesize - dest_filesize
-                    mebibyte = difference/1024/1024
+                    mebibyte = int(difference/1024/1024)
                     logging.info('The new movie is {} MIB smaller than the old one'.format(mebibyte))
 
             # Only copy if checks above passed
